Engine v1, mostly norminette-compliant and documented. Closes #17.
